feat: add tooling for literate tests
This commit adds a new `filecheck` tool which runs tests using a source
file that exemplifies the source file to run a test against. A test file
might look like so:

    # RUN: @midenc compile -o @tempfile @file && cat @tempfile
    # CHECK: <some string that must be in the output>
    ...source code here..

The `@midenc` variable is defined to the absolute path of the
`bin/midenc` artifact, and `@tempfile` will generate a new temporary
file on demand which can be used for output. `@file` refers to the
current file. See the help output of the `filecheck` tool to see how
additional variables can be defined.

By default the test files searched must have the `.mir` extension, but
this can also be configured via arguments to `filecheck`. Running `cargo
make test` will run filecheck tests against files containing both `.mir`
and `.sw` extensions, if they are put in the `tests/lit` directory.

Each test file must specify a `RUN` command, which is run via the current
shell. `CHECK` lines take a regex pattern which must be matched in the output.
`XFAIL` can be used to indicate that the test is expected to fail. Additional markers
are documented in the `lit` crate docs (a dependency of the filecheck tool).