feat: implement stage 1 (hir) infra
The stage 1 IR (or HIR), is a sea-of-nodes style, SSA form intermediate
representation. It is intended to be a target from higher level
languages, so it supports a wider array of operations and control flow
than is supported in MASM (and in the stage 2 IR), but
provides a useful foundation on which to perform the necessary analyses
and transformations in order to lower to MASM.

This commit introduces the basic infrastructure for the IR, namely the
minimum data structures necessary to represent functions in the IR,
which are composed of blocks containing one or more instructions in a
specific order. The set of instructions defined is based on what is
provided in MASM, as well as what is needed to represent programs
written in Sway or Move (barring some features which we can not yet
support in Miden).

Follow on commits will flesh out the IR builder, introduce modules, provide
textual output for debugging, and implement control flow and dominator
tree analyses.