feat: provide type representation enum
The [TypeRepr] enum is intended to represent how a specific [Type] will
be encoded in terms of field elements/words on the operand stack in
Miden. See the doc comments for details.

This is meant to provide a single source for information regarding type
layout in Miden. For types in linear memory, the encoding is slightly
different, as we must preserve the semantics of byte-addressability in
the IR. However, when placing values on the operand stack, the encoding
is modified to provide a more natural alignment between individual terms
and elements on the stack.

The specific representations in this commit are just an initial draft
based on details I'm aware of at this time, but we may make adjustments
as necessary.