fix: reimplement builders for inline asm control flow
The previous implementation did not provide us with an opportunity to
validate the state of the operand stack after the control flow
instruction is constructed. It also failed to account for changes to the
operand stack within the control flow instruction body.

This commit introduces a new implementation, one for if.true, and one
for the loop instructions, which provides a builder API for those
instructions. These builders ensure that upon construction, the operand
stack is consistent across all control flow edges leaving the
instruction, as well as checking other important invariants, such as
while.true requiring a boolean on top of the stack when entering/exiting.