feat: define type compatibility for operators
This implements a function on `Type` called `is_compatible_operand`,
whose purpose is to determine if a type is compatible with another
type in the context of a binary operator. Type compatibility is
defined here in terms of what types can be implicitly coerced to
the controlling type of the operator safely.

For example, one can safely apply `u32.add` to `u32` and `u8`,
where `u32` is the operand whose type "controls" the expected
type produced by the operation. Similarly, applying `u32.add`
to `u8` and `u8` is safe, albeit with no guarantees about catching
overflow of the `u8` range, both operands are `u32` compatible. On the
other hand, applying `u32.add` to `u32` and `i32` is not compatible,
because `u32.add` does not handle signed integers, not to mention there
is no way to handle negative results.

This commit also has a couple small tweaks to how a few of the predicate
functions are expressed to suppress some clippy warnings.