feat: implement support for wasm data segments
Wasm has the concept of "data segments", essentially initializers for
specific regions of memory, identified by offset from the start of
linear memory, and a vector of bytes to be written starting at that
offset. This is used to implement things like the `rodata` section you'd
find in a typical ELF executable/library. Globals can then be exported
with an address pointing into that segment, which can be useful for
sharing the same data across many read-only globals, and for packing
global data more efficiently.

Our implementation of this is very similar, except we also allow data
segments to have a size larger than the initializer (byte vector), which
has the same semantics as padding the initializer with zeroes out to
that size.