feat: implement program linker
This commit introduces [Linker], which is used to take a set of [Module]
and perform many of the tasks that a typical system linker would handle.
In the Miden world, the linker gets involved at a slightly different
stage of compilation however, and does not have the final say in what
code will ultimately get executed at runtime (with some exceptions).

Notably, [Linker] is invoked on a set of modules in IR form, rather than
in Miden Assembly (MASM) form - in other words, the linker is invoked
before modules are lowered to MASM by the codegen backend. This is
because one of the critical tasks performed by the linker, is the
allocation and layout of linear memory for use by the compiled program.
To do this, the linker must determine what set of data segments and
globals will end up in the output, validate that there are no conflicts,
and that all referenced symbols have corresponding definitions, and then
lay out the data segments and globals in memory so that the codegen
backend knows what at what address to find a given global, as well as
where the start of the heap will fall for use by the program.

The [Linker] is also responsible for garbage collecting unused
declarations, but this is dependent on what type of artifact is being
produced.

Since MASM is itself an intermediate representation for the Miden VM,
the compiler does not actually know whether additional modules will be
included in the set provided to the VM at runtime, or even whether all
of the emitted modules will be provided to the VM. That said, we've
designed the linker to make certain assumptions based on what type of
output is being emitted. For example, an executable program is
considered "closed", i.e. we assume that no additional code will be
introduced other than what is being linked (aside from the Miden
standard library). On the other hand, a library program is considered
"open", i.e. we assume that the library will be included as part of a
larger executable program that will contain modules we don't necessarily
control. Despite that, we do make an assumption that any additonal
modules in that situation will play nicely with the memory layout of the
library.