fix: address discrepancies between types and immediates
Types, immediates, and immediate instructions are now uniform, with the
exception that we still have the isize/usize types, but the
corresponding immediates/instructions have been removed in favor of
using more explicit types. In the future we may remove isize/usize types
as well, but for now it may still be useful for things like array
indices and such where the precise type is not important, and the
compiler may choose the most efficient type of its choice.

There are no signed vs unsigned immediates and instructions, matching
recent type system changes.