test: implement multi-module linker test
This commit adds a rather comprehensive test of the linker, as well as a
variety of interesting operations in the IR in general (data segments,
global variables, memory management, etc.)

In the process, I've cleaned up the integration test module a bit, to
make it easier to reuse some of the basic module setup across multiple
tests. No changes to the pre-existing tests aside from moving stuff
around a bit. The linker test is all new.

A number of commits preceding this one implement useful functionality
or fix bugs that came up when writing the linker test.