feat: impelement most of the stackification pass
This commit implements the [Stackify] pass, or at least the high-level
parts of it which are described in its docs. Functions which implement
the low-level details of lowering specific instructions are stubbed out
here to separate those details into a separate commit for review later.

See the [Stackify] docs for a detailed explanation of the pass and its
design/algorithm.

NOTE: This commit represents a work-in-progress state, there may be
things here which are modified as the implementation comes together
across subsequent commits.