refactor: rewrite operand stack impl for codegen
This rewrite makes the operand stack provide a type-agnostic,
operand-oriented API, but internally is type-aware, and ensures
that the layout of operands on the stack is taken into account
when moving things around. It also enhances the representation
of values on the stack to provide richer metadata. In addition,
in order to represent transient values on the stack which do
not have an SSA value, or constant value, associated with them,
we also allow tracking the abstract types of such values on the
stack, while limiting the contexts in which we are allowed to
observe them.