fix: stackify bug in loop exit edge handling
The way I was tracking the "controlling" loop for determining what loop
level an instruction was exiting from was broken. This commit fixes its
deficiencies and places the code related to maintaining that state in a
single place at the entry to `emit`. More tests will be needed to
determine if there are any remaining bugs related to this part of the
stackify pass, but for now this addresses the main bug I was observing.