fix: bug in depgraph node index oracle
This bug was causing instructions to be identified as the last dependent
of a value, thus consuming it, when in fact it was not the last
dependent. This is because the root node in the treegraph was always
being assigned the smallest index, when it should have the largest one.