test: draft whole pipeline compilation and semantic tests
The idea for `CompTest` is to mix and match sources (Rust, Wasm, IR)
and expected compiler output (IR, MASM) to test the whole
pipeline(Rust->Wasm->IR->MASM) or any part of it (Rust->IR, Wasm->IR, IR->MASM, etc).

The compiled MASM code is executed on the emulator and VM and the results
are compared with the native Rust code. The inputs are generated via `proptest`.
The Rust source code consists of the target function under test and an
app binary that calls the target function. The app binary is compiled
to Wasm and then to IR. The target function is called directly in the
test from the linked crate.