feat: add attributes to the ir
Attributes provide a way to decorate IR entities with metadata that
can be used to guide compilation. This is implemented in the parser as
well, but only for functions at this time (since no other entities carry
attributes yet).

In this commit, the only interesting use we care about is identifying
the entrypoint for a program, which can be done by decorating a Function
with the `#[entrypoint]` attribute. Alternatively, one can manually
specify the entrypoint function via Program, assuming it has the
appropriate linkage and calling convention, but the default will be to
search for the `#[entrypoint]` attribute.

This commit also reworks some of the MASM IR structures to account for
how the entrypoint is tracked, as well as where we store the code for
the top-level code block for a program (i.e. where global memory init
will take place, as well as the code responsible for invoking the
user-controllable entrypoint).