feat: revamp cargo-miden to pass the unrecognized options to cargo,
and pick up the artifacts from the cargo build to compile the MASM.
Similar to the way cargo-component does it.

Keep in mind that this cargo extension is planned to be the cargo-component
replacement when it comes to building/publishing/etc the Wasm/Miden components.
This should explain some implementation details. The overaching idea is
to introduce specific to Miden subcommands (e.g. `new`) and pass the
rest of the subcommands to cargo where its (e.g. `build`).