chore: update `cargo-component` to v0.7.0, use `wee_alloc` in Wasm
components and Miden SDK tests.

The panic handler and global allocator is now set in the `lib.rs` file
of each component crate. Before, the `cargo-component-bindings` was not
built correctly for `no_std` and was linking the stdlib allocator.

The Rust bindings for the WIT are now generated in the `src` directory
of the component crate, instead of the `target/.../bindings` directory.

The Wasm `bulk-memory` proposal is enabled via `RUSTFLAGS` so that
`memcpy` function import is replaced with Wasm `memory.copy` op.