feat: implement ilog2/clz/ctz/clo/cto instructions
This implements support for the above-mentioned instructions all the way
through the compiler.

The only thing missing is that we need to update the version of
`miden-assembly` we depend on, which I will be doing in a subsequent PR,
so for now the actual emitting of MASM with these instructions will
fail, but it is trivially fixed after upgrading.

Closes #39