refactor: since all the Miden ABI transformation happens in the frontend
`Abi::Canonical` is used for all the function type definitions because
all of them should have this ABI after the processing
regardless whether an actual transformation was conducted.
This means that `Abi::Miden` is not used, so it can be deleted.
Besides that, I removed `(abi XXXXX)` in `ComponentImport` textual form
and print it as part of `FunctionType` in the textual form.