fix: operand solver improperly tracking aliases
Due to a bug in how the solver context was initialized, expected outputs
with aliases (copies/duplicates) could end up mismatched if the aliases
were at the top of the stack.

This commit addresses that bug, and also extends the CopyInfo struct so
that it can be used for initializing aliases of both the expected
outputs and the current operand stack state.

Closes #200