feat(ir): add support for declaring function-local variables
These locals can be allocated, and then loaded/stored using
local-specific variants of the existing load/store ops. During code
generation, these locals will be lowered to procedure-local memory, and
all loads/stores converted to loads/stores of that memory.

The primary use for locals, is to support alloca-like functionality, as
well as explicit spilling of temporaries to memory during our
pseudo-register allocation pass that determines where spills need to be
inserted so as to keep the operand stack from overflowing 16 elements.