feat(transform): implement spill transform
This commit introduces two passes that are used in conjunction to apply
the result of the spills analysis, introduced previously.

The first pass materializes spills and reloads according to the analysis
results, splitting control flow edges where necessary to accommodate
spill/reload operations along those edges. It updates the spill analysis
results with the concrete blocks and instructions which were
materialized. At this point, the IR is only partially transformed.

The second pass finalizes the transformation, by performing three key
actions:

1. Rewrite all uses of spilled values to use the nearest dominating
   definition. This ensures that uses of a spilled value preceded by a
   reload use the reloaded value, not the original. Furthermore, by
   doing so, this has the effect of splitting the live ranges of all
   spilled values.

2. Allocate sufficient function-local memory to hold spilled values

3. Rewrite all spill/reload ops to explicit stores/loads, making use of
   the allocated function-local memory.

Once complete, the IR of the resulting function is semantically
equivalent to that of the original IR that was analyzed, except it now
additionally has the property that all SSA values that are live at the
same time can be directly fetched from the operand stack using some
sequence of stack ops - this is because we have ensured that none of
those values can end up in the operand stack overflow table, which does
not permit direct access.