test: support tests with non-felt operands
This commit does two things:

* Ensures that tests push operands on the stack in the correct order,
  regardless of how many stack elements each operand requires
* Provides two test-specific traits for encoding/decoding operands
  to/from the operand stack.

This makes it possible to test > 32-bit operands (other than felt), as
previously they weren't supported, or were converting 64-bit operands to
felt, not to the canonical representation of 64-bit values in Miden,
which is two 32-bit limbs.

This also makes it possible to test with non-integral operands as well,
though that is currently untested, and needs to be verified for
correctness.

NOTE: This commit does not update the expect test outputs