wip: refactor compiler test instantiation/configuration
This commit introduces a new `CompilerTestBuilder` type, which is used
to configure compiler tests before they are instantiated. This lets us
build up the configuration that will be used to generate compiler inputs
in multiple stages, and only instantiate it once we wish to run the
test. Prior to this, we mixed these stages together, which made it
difficult to configure the input generation step in more than one place,
which is a problem in more complex tests, or edge case tests which do
not neatly fit the mold of one of the default configurations.

In addition to `CompilerTestBuilder`, there are also two configuration
types introduced, `CargoTest` and `RustcTest`, which configure tests
based on `cargo` or `rustc` outputs (as inputs to `midenc`). These are
used in conjunction with `CompilerTestInputType` to instantiate a
`CompilerTestBuilder` for one of the three input types currently used
for compiler testing.