refactor: unify branch successor repr, support switch arms with arguments
Previously, the Br, CondBr, and Switch branch instructions all used
slightly different representations for their successor information, for
no real purpose. This commit unifies these, as well as provides a more
uniform API (e.g. `analyze_branch`'s `BranchInfo` is also unified in the
same manner).

The Switch instruction type is also enhanced with the ability to pass
block arguments to successor blocks in each arm, as well as the default
case. These statements will be lowered using a binary search approach on
the discriminant value, so as to lower to the least number of
conditional branches possible. Branch weights would be exceedingly
useful here, but we don't have any meaningful way to obtain that
information from Rust currently.

To better support the new Switch API, the `switch` builder has been
updated to use a specialized builder that guarantees that the `Switch`
is well-formed.

The Wasm frontend has had its lowering reworked to make use of the new
builder, and take advantage of the support for successor arguments.