refactor: clean up driver init and output config
This refactoring further simplifies the midenc-compiler interface, and
makes it possible to fully instantiate it from places like cargo-miden
and the test suite.

In addition, the management of output files and directories has been
updated and simplified, and is now a bit more intuitive in terms of how
we derive the path for a given file. There are still some areas to
improve here, but the goal here is to actually make use of the various
flags offered by the CLI.