feat(debugging): improve precision of stack trace generation
This commit uses tracing events attached to `nop` instructions in
function prologue/epilogue to signal when entering/exiting the call
frame of a procedure. Our VM test executor then listens for these trace
events, and records the events at each clock cycle in which they occur.
As we record each instruction executed, we check for events at the same
cycle, and if present, we perform any necessary frame management.p

This commit also refactors/cleans up the test executor, now that a more
precise mechanism is in place for capturing traces. Various improvements
to the output were made as well, so as to make it more useful when
errors occur.