feat(isa): implement dynamic heap memory intrinsics
Prior to this commit, our implementations of `memory_grow` and
`memory_size` were hardcoded to behave as if the heap was already
maximally sized, but various allocators, such as `wee_alloc`, assume
that if they cannot grow the heap, it has already been exhausted, even
though no memory has actually been handed out.

In order to address this, this commit does two things:

1. Implements a set of intrinsics for initializing, reading, and where
   appropriate, mutating, the start, top, and current size of the dynamic
   heap.
2. When creating a [Program], the generated entrypoint initializes the
   heap using information about the layout of linear memory.

Some outdated code in the linker which created a set of predefined
global variables (called HEAP_START, HEAP_TOP, and HEAP_END), have been
removed, as they were not being used, and exposed those globals in the
IR in ways that was not ideal.

Libraries _do not_ initialize the heap, but assume that the program they
link into will do so for them. The intrinsics will assert if that does
not occur, so as to protect programs from silent undefined behavior.