fix: improve quality of wasm translation in a few spots
Due to the way WebAssembly represents its type system, our frontend is
forced to do a lot of pessimistic casts, which result in very poor code
generation without a peephole optimizer in place to clean it up. This
commit tweaks the way we lower certain instructions to reduce the amount
of redundant checks that end up in the generated MASM.

The overall result is not a huge improvement, but helps a bit to reduce
noise. Some thought will need to be put into how we can better mesh the
WebAssembly type system with our own.