feat: implement 'midenc run'
This commit extracts the VM executor and associated helpers from the
miden-integration-tests crate, and moves it into a new midenc-runner
crate, along with the necessary plumbing to wire up `midenc run`.

The actual implementation of the command is stubbed out, but will be
implemented in subsequent commits. The details of how this command will
be used are still up in the air, but for now I plan to use it for
debugging the compiler-emitted MASM programs, by making use of the
debugging infra I've added to our test executor.