feat: implement `miden run` debugger
This commit implements a TUI for an interactive debugger that builds on
top of the previous `midenc-runner` executor and its various diagnostic
tools. See the doc in `midenc-runner/README.md` for usage.

The implementation is based on `ratatui`, and uses a
model/view/controller-like architecture, where the application state is
centrally managed, the top-level `App` structure orchestrates all of the
pages and global events; each page orchestrates its panes and
page-global events; and each pane responds to events, generates its own,
and draws its UI from the current state. This makes it pretty
straightforward to add new pages/panes, keybindings, etc., without it
turning into a gigantic mess.

The initial set of functionality in this commit is actually quite rich,
but there are a few nice-to-haves mentioned in the README, if we can get
to them.