feat(debugger): implement internal debug logging
While working on the debugger, it is occasionally useful to be able to
view debugging information about the debugger program itself, without
resorting to the usual printf tricks, this is a bit tricky. This commit
adds a `debug` command to the debugger, which opens a popup (a new
widget in our UI toolkit) that displays the current stream of events
that have been written to the `log` backend. One can use this to log
something of interest, and then pop upen the debug window to see the
event in question.

It is currently a bit noisy, with no way to meaningfully filter the log
yet, but for now it is sufficient, and servces as a useful example of
the popup functionality that we can use for other things as well.