feat: implement packaging prototype
This commit implements a basic package format for Miden programs and
libraries, with the metadata needed to simplify using them together. The
format is experimental, and can be changed at any time, however it does
use a header that lets us identify whether a package was produced with
the same specification or not.

This also modifies the compiler and debugger to emit and consume,
respectively, the package format as their primary artifact. The debugger
uses this, for example, to initialize the VM with the necessary
libraries and rodata segments required by the program.

The package format is described using the structures in
`codegen/masm/src/packaging/package.rs`, and is encoded using a tiny,
but efficient binary format from the `bitcode` crate, using `serde` to
facilitate the lowering of our high-level types to that format. This
made it relatively straightforward to implement the format without
having to mess around with the low-level byte representation. The
`bitcode` crate is designed with stability in mind, though because we
are using it via `serde`, the stability is subject to being violated if
`serde` changes something that causes `bitcode` to serialize a type
differently. For now, this is a non-issue, longer term we might want to
actually implement `bitcode::{Encode, Decode}` for our types if we want
to gain more stability.