fix(frontend-wasm): reserve memory allocated for use by rust
It turns out we were assuming that Rust was laying out its memory using
the default LLVM layout, but rustc actually specifies a larger shadow
stack. We were not properly handling the defined memory in Wasm modules
parsed by the frontend, and so we ended up incorrectly placing global
variables and other data in the middle of memory that rust was using for
either the shadow stack or static data. Because of the size of the
shadow stack region reserved by Rust, this was not causing any
corruption in our tests, but it was causing other assumptions to be
invalid, and miscompilations were the result.

This commit treats the Wasm memory for a parsed module as reserved for
use by the code in that module, which causes midenc to place any global
variables it defines after that region, and computes the start of the
heap as immediately following any such data. It does so by propagating
the page size and reserved number of pages through the IR and to the
backend, taking that information into account as needed.