ci: rework github actions workflows
* Replace unmaintained actions-rs steps with manual steps
* Rely on rust-toolchain.toml file for installing required toolchain,
  targets and components. This also cleans up Makefile.toml a bit
* Improve the way we handle build caches using Swatinem/rust-cache. See
  below for details.
* Restructure jobs to run lint/formatter checks and tests in parallel,
  but introduce some ordering in those groups to improve cache coherency
  and reuse, and in some cases, to avoid running expensive actions which
  are very likely to fail anyway. The main difference here is that
  amongst the test group, the unit test job is run first, followed by
  the midenc and cargo-miden integration test jobs. Previously only the
  cargo-miden integration tests were split out into a separate job. By
  breaking apart unit and integration tests, we can fail fast when unit
  tests are broken, and avoid running the expensive integration tests
  when they are very likely to fail if the unit tests are broken.
* Update the `book.yml` workflow to publish our new mkdocs-based
  documentation, rather than the old mdbook-based docs. This was causing
  the workflow to fail, it should now start working again.

RE: caching in CI. See https://github.com/Swatinem/rust-cache for
details on how the caching step works, particularly _what_ is cached,
how it is keyed, and _when_ it is cached.

The changes in this commit cache the results of a successful run _only_
when a push to the 'next' branch occurs, i.e. when a PR is merged. Thus,
all PR runs will use the cache associated with the last successful build
of `next` as a base, and build off of that.

Furthermore, there are three different caches:

* The cache shared by non-test jobs in the `ci.yml` workflow
* The cache shared by test jobs in the `ci.yml` workflow
* The cache used by the release job in the `release.yml` workflow

Previous to this commit, each job had its own cache, and the cache would
be updated on every successful build, causing a fair amount of cache
instability, and resulting in very little reuse.

This commit is more conservative about updating the cache, but shares it
more broadly across jobs, and the more conservative baseline means that
the cache is hit more frequently across all PRs.

The choice to not share the cache between non-test and test jobs in the
`ci.yml` workflow is because we want to maximize the number of jobs that
run in parallel, but to reuse the cache, we must ensure that only one
job is responsible for maintaining it. If we maximized cache reuse, we'd
run all jobs in sequence, but this would be slow. Instead, we break up
the workflow into non-test checks, and tests, and pick a baseline job in
both groups to handle maintaining the cache for that group. These groups
can then be run in parallel, and after the job that maintains the cache
is run, all remaining jobs in that group can be run in parallel.