wip: implement key infrastructure for operations, patterns, builders
This commit moves away from `derive!` for operation definitions, to the
new `#[operation]` macro, and implements a number of
changes/improvements to the `Operation` type and its related APIs.

In particular with this commit, the builder infrastructure for
operations was finalized and started to be tested with "real" ops.
Validation was further improved by typing operands with type constraints
that are validated when an op is verified. The handling of successors,
operands, and results was improved and unified around a shared
abstraction.

The pattern and pattern rewriter infrastructure was sketched out as
well, and I anticipate landing those next, along with tests for all of
it.

Once patterns are done, the next two items of note are the data analysis
framework, and the conversion framework. With those done, we're ready to
move to the new IR.